+++ src/components/BaseTooltipTitle.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from "react";
 import { Avatar, makeStyles, Box, Typography } from "@material-ui/core";
-import { theme } from "@/theme";
-import { ToolTip } from "@/tooltips";
+import { theme } from "@/components/theme";
+import { ToolTip } from "@/components/tooltips";
 
 const useStyles = (props: Props) =>
   makeStyles({
