+++ src/components/palette.ts
